create SoftwareCommands.{c,h}
Move a bunch more commands out of CommandLineTask.c
